Module imageproc::affine

Functions for affine transformations of images.

Enums

Interpolation
How to handle pixels whose pre-image lies between input pixels.

Functions

affine
Applies an affine transformation to an image, or None if the provided transformation is not invertible. The output image has the same dimensions as the input. Output pixels whose pre-image lies outside the input image are set to black.

affine_with_default
Applies an affine transformation to an image, or None if the provided transformation is not invertible. The output image has the same dimensions as the input. Output pixels whose pre-image lies outside the input image are set to default.

rotate
Rotate an image clockwise about provided center by theta radians. The output image has the same dimensions as the input. Output pixels whose pre-image lies outside the input image are black.

rotate_about_center
Rotate an image clockwise about its center by theta radians. The output image has the same dimensions as the input. Output pixels whose pre-image lies outside the input image are black.

rotate_with_default
Rotate an image clockwise about provided center by theta radians. The output image has the same dimensions as the input. Output pixels whose pre-image lies outside the input image are set to default.

translate
Translates the input image by t. Note that image coordinates increase from top left to bottom right. Output pixels whose pre-image are not in the input image are set to the boundary pixel in the input image nearest to their pre-image.