Function imageproc::affine::affine_with_default

pub fn affine_with_default<I>(image: &I, affine: Affine2, default: I::Pixel, interpolation: Interpolation) -> Option<VecBuffer<I::Pixel>> where I: GenericImage, I::Pixel: 'static, I::Pixel::Subpixel: ValueInto<f32> + Clamp<f32>

Applies an affine transformation to an image, or None if the provided transformation is not invertible. The output image has the same dimensions as the input. Output pixels whose pre-image lies outside the input image are set to default.