Module imageproc::definitions

Trait definitions and type aliases.

Traits

Clamp
A type to which we can clamp a value of type T. Implementations are not required to handle NaNs gracefully.

HasBlack
Pixels which have a named Black value.

HasWhite
Pixels which have a named White value.

Position
Something with a 2d position.

Score
Something with a score.

Type Definitions

VecBuffer
An ImageBuffer containing Pixels of type P with storage Vec.