Enum imageproc::distance_transform::Norm

pub enum Norm {
    L1,
    LInf,
}

How to measure distance between coordinates. See the distance_transform documentation for examples.

Note that this enum doesn't currently include the L2 norm. As Norm is used by the morphology functions, this means that we don't support using the L2 norm for any of those functions.

This module does support calculating the L2 distance function, via the euclidean_squared_distance_transform function, but the signature of this function is not currently compatible with those for computing L1 and LInf distance transforms. It would be nice to unify these functions in future.

Variants
L1

Defines d((x1, y1), (x2, y2)) to be abs(x1 - x2) + abs(y1 - y2). Also known as the Manhattan or city block norm.

LInf

Defines d((x1, y1), (x2, y2)) to be max(abs(x1 - x2), abs(y1 - y2)). Also known as the chessboard norm.