Function imageproc::drawing::draw_antialiased_line_segment

pub fn draw_antialiased_line_segment<I, B>(
    image: &I, 
    start: (i32, i32), 
    end: (i32, i32), 
    color: I::Pixel, 
    blend: B
) -> Image<I::Pixel> where
    I: GenericImage,
    I::Pixel: 'static,
    B: Fn(I::Pixel, I::Pixel, f32) -> I::Pixel, 

Draws as much of the line segment between start and end as lies inside the image bounds. The parameters of blend are (line color, original color, line weight). Consider using interpolate for blend. Uses Xu's line drawing algorithm.