Function imageproc::drawing::draw_text_mut

pub fn draw_text_mut<'a, C>(
    canvas: &'a mut C, 
    color: C::Pixel, 
    x: u32, 
    y: u32, 
    scale: Scale, 
    font: &'a Font<'a>, 
    text: &'a str
) where
    C: Canvas,
    <C::Pixel as Pixel>::Subpixel: ValueInto<f32> + Clamp<f32>, 

Draws colored text on an image in place. scale is augmented font scaling on both the x and y axis (in pixels). Note that this function does not support newlines, you must do this manually