Function imageproc::edges::canny

pub fn canny<I>(image: &I,
                low_threshold: f32,
                high_threshold: f32)
                -> ImageBuffer<Luma<u8>, Vec<u8>> where I: GenericImage<Pixel=Luma<u8>> + 'static

Runs the canny edge detection algorithm on the provided ImageBuffer.

Params

low_threshold: Low threshold for the hysteresis procedure. Edges with a strength higher than the low threshold will appear in the output image, if there are strong edges nearby.

high_threshold: High threshold for the hysteresis procedure. Edges with a strength higher than the high threshold will always appear as edges in the output image.

Returns a binary image, where edge pixels have a value of 255 and non-edge pixels a value of 0.