Struct imageproc::filter::Kernel

pub struct Kernel<'a, K: 'a> { /* fields omitted */ }

A 2D kernel, used to filter images via convolution.

Methods

impl<'a, K: Num + Copy + 'a> Kernel<'a, K>

fn new(data: &'a [K], width: u32, height: u32) -> Kernel<'a, K>

Construct a kernel from a slice and its dimensions. The input slice is in row-major form.

fn filter<I, F, Q>(&self, image: &I, f: F) -> VecBuffer<Q> where I: GenericImage,
I::Pixel::Subpixel: ValueInto<K>,
Q: Pixel + 'static,
F: FnMut(&mut Q::Subpixel, K) -> ()

Returns 2d correlation of an image. Intermediate calculations are performed at type K, and the results converted to pixel Q via f. Pads by continuity.