Module imageproc::haar

Functions for creating and evaluating Haar-like features.

Structs

HaarFilter
A Haar filter whose value on an integral image is the weighted sum of the values of the integral image at the given points.

Enums

Sign
Whether the top left region in a Haar filter is counted with positive or negative sign.

Functions

draw_haar_filter
Draws the given Haar filter on an image, drawing pixels with a positive sign white and those with a negative sign black.

draw_haar_filter_mut
Draws the given Haar filter on an image in place, drawing pixels with a positive sign white and those with a negative sign black.

enumerate_haar_filters
Returns a vector of all valid Haar filters for an image with given width and height.