Enum imageproc::haar::HaarFeatureType

pub enum HaarFeatureType {
    TwoRegionHorizontal,
    ThreeRegionHorizontal,
    TwoRegionVertical,
    ThreeRegionVertical,
    FourRegion,
}

The type of a Haar-like feature determines the number of regions it contains and their orientation. The diagrams in the comments for each variant use the symbols (*, &) to represent either (+, -) or (-, +), depending on which Sign the feature type is used with.

Variants
TwoRegionHorizontal

Two horizontally-adjacent regions of equal width.

     -----------
    |  *  |  &  |
     -----------

ThreeRegionHorizontal

Three horizontally-adjacent regions of equal width.

     -----------------
    |  *  |  &  |  *  |
     -----------------

TwoRegionVertical

Two vertically-adjacent regions of equal height.

     -----
    |  *  |
     -----
    |  &  |
     -----

ThreeRegionVertical

Three vertically-adjacent regions of equal height.

     -----
    |  *  |
     -----
    |  &  |
     -----
    |  *  |
     -----

FourRegion

Four regions arranged in a two-by-two grid. The two columns have equal width and the two rows have equal height.

     -----------
    |  *  |  &  |
     -----------
    |  &  |  *  |
     -----------