Function imageproc::haar::draw_haar_filter_mut

pub fn draw_haar_filter_mut<I>(image: &mut I, filter: HaarFilter) where I: GenericImage, I::Pixel: HasBlack + HasWhite

Draws the given Haar filter on an image in place, drawing pixels with a positive sign white and those with a negative sign black.