Function imageproc::haar::number_of_haar_filters

pub fn number_of_haar_filters(width: u32, height: u32) -> u32

Returns the number of distinct Haar filters for an image of the given dimensions. Includes positive and negative, two and three region, vertical and horizontal filters, as well as positive and negative four region filters.

Consider two-region positive horizontal Haar filters in an image of height 1. There is exactly one such filter for each choice of L, M and R below.

    L   M     R
| | | | | | | | |
 . . + + - - - .

There are (width + 1) dividing lines between pixels, so ((width + 1) choose 3) such filters. For an image of arbitrary height there are ((height + 1) choose 2) choices for the top and bottom of each two region positive Haar filter, and every positive filter has a corresponding negative filter. Thus there are ((width + 1) choose 3) * ((height + 1) choose 2) * 2 filters of this type in an image with dimensions (width, height).

A very similar argument applies to the other filter types.