Struct imageproc::haar::HaarFilter

pub struct HaarFilter {
    // some fields omitted
}

A Haar filter whose value on an integral image is the weighted sum of the values of the integral image at the given points.

Methods

impl HaarFilter

fn evaluate<I>(&self, integral: &I) -> i32 where I: GenericImage<Pixel=Luma<u32>>

Evaluates the Haar filter on an integral image.

fn two_region_horizontal(top: u32, left: u32, dx1: u32, dx2: u32, dy: u32, sign: Sign) -> HaarFilter

Returns the following feature (with signs reversed if Sign == Sign::Negative).

    A   B   C
      +   -
    D   E   F

A = (top, left), B.x = left + dx1, C.x = B.x + dx2, and D.y = A.y + dy.

fn two_region_vertical(top: u32, left: u32, dx: u32, dy1: u32, dy2: u32, sign: Sign) -> HaarFilter

Returns the following feature (with signs reversed if Sign == Sign::Negative).

    A   B
      +
    C   D
      -
    E   F

A = (top, left), B.x = left + dx, C.y = top + dy1, and E.y = C.y + dy2.

fn three_region_horizontal(top: u32, left: u32, dx1: u32, dx2: u32, dx3: u32, dy: u32, sign: Sign) -> HaarFilter

Returns the following feature (with signs reversed if Sign == Sign::Negative).

    A   B   C   D
      +   -   +
    E   F   G   H

A = (top, left), B.x = left + dx1, C.x = B.x + dx2, D.x = C.x + dx3, and E.y = top + dy.

fn three_region_vertical(top: u32, left: u32, dx: u32, dy1: u32, dy2: u32, dy3: u32, sign: Sign) -> HaarFilter

Returns the following feature (with signs reversed if Sign == Sign::Negative).

    A   B
      +
    C   D
      -
    E   F
      +
    G   H

A = (top, left), B.x = left + dx, C.y = top + dy1, E.y = C.y + dy2, and G.y = E.y + dy3.

fn four_region(top: u32, left: u32, dx1: u32, dx2: u32, dy1: u32, dy2: u32, sign: Sign) -> HaarFilter

Returns the following feature (with signs reversed if Sign == Sign::Negative).

    A   B   C
      +   -
    D   E   F
      -   +
    G   H   I

A = (top, left), B.x = left + dx1, C.x = B.x + dx2, D.y = top + dy1, and G.y = D.y + dy2.