Module imageproc::hough

Line detection via the Hough transform.

Structs

LineDetectionOptions
Options for Hough line detection.

PolarLine
A detected line, in polar coordinates.

Functions

detect_lines
Detects lines in a binary input image using the Hough transform.

draw_polar_lines
Draws each element of lines on image in the provided color.

draw_polar_lines_mut
Draws each element of lines on image in the provided color.