Function imageproc::hough::detect_lines

pub fn detect_lines(
    image: &GrayImage, 
    options: LineDetectionOptions
) -> Vec<PolarLine>

Detects lines in a binary input image using the Hough transform.

Points are considered to be in the foreground (and thus vote for lines) if their intensity is non-zero.

See ./examples/hough.rs for example usage.