Module imageproc::local_binary_patterns

Functions for computing local binary patterns.

Statics

MIN_SHIFT
Lookup table for the least circular shift of a byte.

UNIFORM_REPRESENTATIVE_2
Maps uniform bytes (i.e. those with at most two bit transitions) to their least circular shifts, and non-uniform bytes to 10101010 (an arbitrarily chosen non-uniform representative).

Functions

count_transitions
Number of bit transitions in a byte, counting the last and final bits as adjacent.

local_binary_pattern
Computes the basic local binary pattern of a pixel, or None if it's too close to the image boundary.

min_shift
Returns the least value of all rotations of a byte.