Module imageproc::map

Functions for mapping over pixels, colors or subpixels of images.

Traits

WithChannel
The type obtained by replacing the channel type of a given Pixel type.

Functions

as_blue_channel
Create an RGB image by embedding a grayscale image in a single channel.

as_green_channel
Create an RGB image by embedding a grayscale image in a single channel.

as_red_channel
Create an RGB image by embedding a grayscale image in a single channel.

blue_channel
Create a grayscale image by extracting a channel of an RGB image.

green_channel
Create a grayscale image by extracting a channel of an RGB image.

map_colors
Applies f to the color of each pixel in the input image.

map_pixels
Applies f to each pixel in the input image.

map_subpixels
Applies f to each subpixel of the input image.

red_channel
Create a grayscale image by extracting a channel of an RGB image.

Type Definitions

ChannelMap
Alias to make uses of WithChannel less syntactically noisy.