Trait imageproc::map::WithChannel

pub trait WithChannel<C: Primitive>: Pixel {
    type Pixel: Pixel<Subpixel = C> + 'static;
}

The type obtained by replacing the channel type of a given Pixel type. The output type must have the same name of channels as the input type, or several algorithms will produce incorrect results or panic.

Associated Types

type Pixel: Pixel<Subpixel = C> + 'static

The new pixel type.

Loading content...