Struct imageproc::math::Vec2

#[repr(C)]

pub struct Vec2<N> {
    pub x: N,
    pub y: N,
}

Vector of dimension 2.

Fields
x: N

First component of the vector.

y: N

Second component of the vector.

Methods

impl<N> Vec2<N>

fn new(x: N, y: N) -> Vec2<N>

Creates a new vector.

impl<N> Vec2<N> where N: Zero + One

fn x() -> Vec2<N>

Create a unit vector with its $compN component equal to 1.0.

fn y() -> Vec2<N>

Create a unit vector with its $compN component equal to 1.0.

impl<N> Vec2<N> where N: Copy

unsafe fn at_fast(&self, i: usize) -> N

Unsafe read access to a vector element by index.

unsafe fn set_fast(&mut self, i: usize, val: N)

Unsafe write access to a vector element by index.

impl<N> Vec2<N>

fn len(&self) -> usize

impl<N> Vec2<N>

fn to_pnt(self) -> Pnt2<N>

fn as_pnt(&self) -> &Pnt2<N>