Function imageproc::morphology::distance_transform_mut

pub fn distance_transform_mut(image: &mut GrayImage, norm: Norm)

Updates an image in place so that each pixel contains its distance from a foreground pixel in the original image.

A pixel belongs to the foreground if it has non-zero intensity. As the image has a bit-depth of 8, distances saturate at 255.

See the distance_transform documentation for examples.