Struct imageproc::multiarray::View3d

pub struct View3d<'a, T: 'a> {
    pub data: &'a mut [T],
    pub lengths: [usize; 3],
}

A view into a 3d array.

Fields

data

THe underlying data.

lengths

Lengths of the dimensions, from innermost (i.e. fastest-varying) to outermost.

Methods

impl<'a, T> View3d<'a, T>

fn from_raw(data: &'a mut [T], lengths: [usize; 3]) -> View3d<'a, T>

Constructs index from existing data and the lengths of the desired dimensions.

fn data(&self) -> &[T]

Immutable access to the raw data.

fn data_mut(&mut self) -> &mut [T]

Mutable access to the raw data.

fn at(&self, indices: [usize; 3]) -> &T

An immutable reference from a 3d index.

fn at_mut(&mut self, indices: [usize; 3]) -> &mut T

A mutable reference from a 3d index.

fn inner_slice(&self, x1: usize, x2: usize) -> &[T]

All entries with the given outer dimensions. As the first dimension is fastest varying, this is a contiguous slice.

fn inner_slice_mut(&mut self, x1: usize, x2: usize) -> &mut [T]

All entries with the given outer dimensions. As the first dimension is fastest varying, this is a contiguous slice.