Struct imageproc::rect::Rect

pub struct Rect { /* fields omitted */ }

A rectangular region of non-zero width and height.

Methods

impl Rect

fn at(x: i32, y: i32) -> RectPosition

Reduces possibility of confusing coordinates and dimensions when specifying rects.

fn top(&self) -> i32

Smallest y-coordinate reached by rect.

fn left(&self) -> i32

Smallest x-coordinate reached by rect.

fn bottom(&self) -> i32

Greatest y-coordinate reached by rect.

fn right(&self) -> i32

Greatest x-coordinate reached by rect.

fn width(&self) -> u32

Width of rect.

fn height(&self) -> u32

Height of rect.

fn intersect(&self, other: Rect) -> Option<Rect>

Returns the intersection of self and other, or none if they are are disjoint.