Module imageproc::seam_carving

An implementation of seam carving. Currently in a pretty rough state. See examples/seam_carving.rs for an example.

Structs

VerticalSeam
An image seam connecting the bottom of an image to its top (in that order).

Functions

draw_vertical_seams
Draws a series of seams on image in red. Assumes that the provided seams were removed in the given order from the input image.

find_vertical_seam
Computes an 8-connected path from the bottom of the image to the top whose sum of gradient magnitudes is minimal.

remove_vertical_seam
Returns the result of removing seam from image.

shrink_width
Reduces the width of an image using seam carving.