Function imageproc::seam_carving::find_vertical_seam

pub fn find_vertical_seam<P>(image: &Image<P>) -> VerticalSeam where
    P: Pixel<Subpixel = u8> + WithChannel<u16> + WithChannel<i16> + 'static,
    <P as WithChannel<u16>>::Pixel: HasBlack, 

Computes an 8-connected path from the bottom of the image to the top whose sum of gradient magnitudes is minimal.