Module imageproc::stats

Statistical properties of images.

Functions

cumulative_histogram
Returns the cumulative histogram of grayscale values in an 8bpp grayscale image.

histogram
Returns the histogram of grayscale values in an 8bpp grayscale image.

peak_signal_to_noise_ratio
Returns the peak signal to noise ratio for a clean image and its noisy aproximation. All channels are considered equally. If you do not want this (e.g. if using RGBA) then change image formats first. See also peak signal-to-noise ratio (wikipedia).

percentile
Returns the pth percentile of the pixel intensities in an image.

root_mean_squared_error
Returns the square root of the mean of the squares of differences between all subpixels in left and right. All channels are considered equally. If you do not want this (e.g. if using RGBA) then change image formats first.