Struct imageproc::union_find::DisjointSetForest

pub struct DisjointSetForest { /* fields omitted */ }

Data structure for efficient union find.

Methods

impl DisjointSetForest

pub fn new(count: usize) -> DisjointSetForest

Constructs forest of singletons with count elements.

pub fn num_trees(&self) -> usize

Returns the number of trees in the forest.

pub fn root(&mut self, i: usize) -> usize

Returns index of the root of the tree containing i. Needs mutable reference to self for path compression.

pub fn find(&mut self, i: usize, j: usize) -> bool

Returns true if i and j are in the same tree. Need mutable reference to self for path compression.

pub fn union(&mut self, i: usize, j: usize)

Unions the trees containing i and j.

pub fn trees(&mut self) -> Vec<Vec<usize>>

Returns the elements of each tree.