Struct imageproc::unionfind::DisjointSetForest

pub struct DisjointSetForest {
    // some fields omitted
}

Data structure for efficient union find.

Methods

impl DisjointSetForest

fn new(count: usize) -> DisjointSetForest

Constructs forest of singletons with count elements.

fn root(&mut self, i: usize) -> usize

Returns index of the root of the tree containing i. Needs mutable reference to self for path compression.

fn find(&mut self, i: usize, j: usize) -> bool

Returns true if i and j are in the same tree. Need mutable reference to self for path compression.

fn union(&mut self, i: usize, j: usize)

Unions the trees containing i and j.

fn trees(&mut self) -> Vec<Vec<usize>>

Returns the elements of each tree.