Crate imageproc

Modules

affine
Functions for affine transformations of images.

contrast
Functions for manipulating the contrast of images.

corners
Functions for detecting corners, also known as interest points.

definitions
Trait definitions and type aliases.

drawing
Helpers for drawing basic shapes on images.

edges
Functions for detecting edges in images.

filter
Functions for filtering images.

gradients
Functions for computing gradients of image intensities.

haar
Functions for creating and evaluating Haar-like features.

hog
HoG features and helpers for visualizing them.

integralimage
Functions for computing integral images and running sums of rows and columns.

localbinarypatterns
Functions for computing local binary patterns.

map
Functions for mapping over pixels, colors or subpixels of images.

math
Assorted mathematical helper functions.

multiarray
Helpers providing multi-dimensional views of 1d data.

noise
Functions for adding synthetic noise to images.

pixelops
Pixel manipulations.

rect
Basic manipulation of rectangles.

regionlabelling
Functions for finding and labelling connected components of an image.

stats
Statistical properties of images.

suppress
Functions for suppressing non-maximal values.

unionfind
An implementation of disjoint set forests for union find.

utils
Utils for testing and debugging.

Macros

assert_dimensions_match
Panics if image dimensions do not match.

assert_pixels_eq
Panics if any pixels differ between the two input images.

assert_pixels_eq_within
Panics if any pixels differ between the two images by more than the given tolerance in a single channel.