Crate imageproc

An image processing library, based on the image crate.

Modules

affine
Functions for affine transformations of images.

contrast
Functions for manipulating the contrast of images.

corners
Functions for detecting corners, also known as interest points.

definitions
Trait definitions and type aliases.

drawing
Helpers for drawing basic shapes on images.

edges
Functions for detecting edges in images.

filter
Functions for filtering images.

gradients
Functions for computing gradients of image intensities.

haar
Functions for creating and evaluating Haar-like features.

hog
HoG features and helpers for visualizing them.

hough
Line detection via the Hough transform.

integral_image
Functions for computing integral images and running sums of rows and columns.

local_binary_patterns
Functions for computing local binary patterns.

map
Functions for mapping over pixels, colors or subpixels of images.

math
Assorted mathematical helper functions.

morphology
Functions for computing morphological operators.

noise
Functions for adding synthetic noise to images.

pixelops
Pixel manipulations.

property_testing
Utilities to help with writing property-based tests (e.g. quickcheck tests) for image processing functions.

rect
Basic manipulation of rectangles.

region_labelling
Functions for finding and labelling connected components of an image.

seam_carving
An implementation of seam carving. Currently in a pretty rough state. See examples/seam_carving.rs for an example.

stats
Statistical properties of images.

suppress
Functions for suppressing non-maximal values.

union_find
An implementation of disjoint set forests for union find.

utils
Utils for testing and debugging.

Macros

assert_dimensions_match
Panics if image dimensions do not match.

assert_pixels_eq
Panics if any pixels differ between the two input images.

assert_pixels_eq_within
Panics if any pixels differ between the two images by more than the given tolerance in a single channel.

gray_image
Helper for defining greyscale images.

rgb_image
Helper for defining RGB images.