Struct imageproc::filter::Kernel

pub struct Kernel<'a, K: 'a> { /* fields omitted */ }

A 2D kernel, used to filter images via convolution.

Methods

impl<'a, K: Num + Copy + 'a> Kernel<'a, K>

pub fn new(data: &'a [K], width: u32, height: u32) -> Kernel<'a, K>

Construct a kernel from a slice and its dimensions. The input slice is in row-major form.

pub fn filter<P, F, Q>(&self, image: &Image<P>, f: F) -> Image<Q> where
P: Pixel + 'static,
<P as Pixel>::Subpixel: ValueInto<K>,
Q: Pixel + 'static,
F: FnMut(&mut Q::Subpixel, K),

Returns 2d correlation of an image. Intermediate calculations are performed at type K, and the results converted to pixel Q via f. Pads by continuity.