Module imageproc::gradients

Functions for computing gradients of image intensities.

Statics

HORIZONTAL_PREWITT
Prewitt filter for detecting horizontal gradients.

HORIZONTAL_SOBEL
Sobel filter for detecting horizontal gradients.

VERTICAL_PREWITT
Prewitt filter for detecting vertical gradients.

VERTICAL_SOBEL
Sobel filter for detecting vertical gradients.

Functions

horizontal_prewitt
Convolves an image with the HORIZONTAL_PREWITT kernel to detect horizontal gradients.

horizontal_sobel
Convolves an image with the HORIZONTAL_SOBEL kernel to detect horizontal gradients.

prewitt_gradients
Returns the magnitudes of gradients in an image using Prewitt filters.

sobel_gradient_map
Computes per-channel gradients using Sobel filters and calls f to compute each output pixel.

sobel_gradients
Returns the magnitudes of gradients in an image using Sobel filters.

vertical_prewitt
Convolves an image with the VERTICAL_PREWITT kernel to detect vertical gradients.

vertical_sobel
Convolves an image with the VERTICAL_SOBEL kernel to detect vertical gradients.