Struct imageproc::hog::HogOptions

pub struct HogOptions {
    pub orientations: usize,
    pub signed: bool,
    pub cell_side: usize,
    pub block_side: usize,
    pub block_stride: usize,
}

Parameters for HoG descriptors.

Fields

orientations

Number of gradient orientation bins.

signed

Whether gradients in opposite directions are treated as equal.

cell_side

Width and height of cell in pixels.

block_side

Width and height of block in cells.

block_stride

Offset of the start of one block from the next in cells.

Methods

impl HogOptions

fn new(orientations: usize, signed: bool, cell_side: usize, block_side: usize, block_stride: usize) -> HogOptions