Module imageproc::integral_image

Functions for computing integral images and running sums of rows and columns.

Functions

column_running_sum
Computes the running sum of one column of image, padded at the top and bottom. The padding is by continuity. Takes a reference to buffer so that this can be reused for all columns in an image.

integral_image
Computes the 2d running sum of a grayscale image.

integral_squared_image
Computes the 2d running sum of the squares of the intensities in a grayscale image.

row_running_sum
Computes the running sum of one row of image, padded at the beginning and end. The padding is by continuity. Takes a reference to buffer so that this can be reused for all rows in an image.

sum_image_pixels
Sums the pixels in positions [left, right] * [top, bottom] in F, where integral_image is the integral image of F.

variance
Computes the variance of [left, right] * [top, bottom] in F, where integral_image is the integral image of F and integral_squared_image is the integral image of the squares of the pixels in F.