Module imageproc::noise

Functions for adding synthetic noise to images.

Functions

gaussian_noise
Adds independent additive Gaussian noise to all channels of an image, with the given mean and standard deviation.

gaussian_noise_mut
Adds independent additive Gaussian noise to all channels of an image in place, with the given mean and standard deviation.

salt_and_pepper_noise
Converts pixels to black or white at the given rate (between 0.0 and 1.0). Black and white occur with equal probability.

salt_and_pepper_noise_mut
Converts pixels to black or white in place at the given rate (between 0.0 and 1.0). Black and white occur with equal probability.