Struct imageproc::rect::Rect

pub struct Rect { /* fields omitted */ }

A rectangular region of non-zero width and height.

Examples

use imageproc::rect::Rect;
use imageproc::rect::Region;

// Construct a rectangle with top-left corner at (4, 5), width 6 and height 7.
let rect = Rect::at(4, 5).of_size(6, 7);

// Contains top-left point:
assert_eq!(rect.left(), 4);
assert_eq!(rect.top(), 5);
assert!(rect.contains(rect.left(), rect.top()));

// Contains bottom-right point, at (left + width - 1, top + height - 1):
assert_eq!(rect.right(), 9);
assert_eq!(rect.bottom(), 11);
assert!(rect.contains(rect.right(), rect.bottom()));

Methods

impl Rect

pub fn at(x: i32, y: i32) -> RectPosition

Reduces possibility of confusing coordinates and dimensions when specifying rects.

See the struct-level documentation for examples.

pub fn top(&self) -> i32

Smallest y-coordinate reached by rect.

See the struct-level documentation for examples.

pub fn left(&self) -> i32

Smallest x-coordinate reached by rect.

See the struct-level documentation for examples.

pub fn bottom(&self) -> i32

Greatest y-coordinate reached by rect.

See the struct-level documentation for examples.

pub fn right(&self) -> i32

Greatest x-coordinate reached by rect.

See the struct-level documentation for examples.

pub fn width(&self) -> u32

Width of rect.

pub fn height(&self) -> u32

Height of rect.

pub fn intersect(&self, other: Rect) -> Option<Rect>

Returns the intersection of self and other, or none if they are are disjoint.

Examples

use imageproc::rect::Rect;
use imageproc::rect::Region;

// Intersecting a rectangle with itself
let r = Rect::at(4, 5).of_size(6, 7);
assert_eq!(r.intersect(r), Some(r));

// Intersecting overlapping but non-equal rectangles
let r = Rect::at(0, 0).of_size(5, 5);
let s = Rect::at(1, 4).of_size(10, 12);
let i = Rect::at(1, 4).of_size(4, 1);
assert_eq!(r.intersect(s), Some(i));

// Intersecting disjoint rectangles
let r = Rect::at(0, 0).of_size(5, 5);
let s = Rect::at(10, 10).of_size(100, 12);
assert_eq!(r.intersect(s), None);